+++ test-templates/src/msm.rs
@@ -1,4 +1,7 @@
-use ark_ec::{msm::VariableBase, AffineCurve, ProjectiveCurve};
+use ark_ec::{
+    msm::{ChunkedPippenger, VariableBase},
+    AffineCurve, ProjectiveCurve,
+};
 use ark_ff::{PrimeField, UniformRand, Zero};
 
 fn naive_var_base_msm<G: AffineCurve>(
@@ -30,4 +33,27 @@ pub fn test_var_base_msm<G: AffineCurve>() {
     let fast = VariableBase::msm(g.as_slice(), v.as_slice());
 
     assert_eq!(naive.into_affine(), fast.into_affine());
+}
+
+pub fn test_chunked_pippenger<G: AffineCurve>() {
+    const SAMPLES: usize = 1 << 10;
+
+    let mut rng = ark_std::test_rng();
+
+    let v = (0..SAMPLES)
+        .map(|_| G::ScalarField::rand(&mut rng).into_repr())
+        .collect::<Vec<_>>();
+    let g = (0..SAMPLES)
+        .map(|_| G::Projective::rand(&mut rng))
+        .collect::<Vec<_>>();
+    let g = <G::Projective as ProjectiveCurve>::batch_normalization_into_affine(&g);
+
+    let arkworks = VariableBase::msm(g.as_slice(), v.as_slice());
+
+    let mut p = ChunkedPippenger::<G>::new(1 << 20);
+    for (s, g) in v.iter().zip(g) {
+        p.add(g, s);
+    }
+    let mine = p.finalize();
+    assert_eq!(arkworks.into_affine(), mine.into_affine());
 }